+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2.5-dev - Unreleased
 - Fix month-day formatting ("January 1st" -> "January 1")
 - Change ordinal formatting to use words instead of numerals for 1-9
+- Fix digital time formatting (use HH:MM instead H:MM)
 
 ## 2.4 - 2023-04-26
 - Add explain() method which describes the expression in human language